Break out matcher/re.hpp
